Themes: Handle case when keyguard is showing during theme change
After we recreate the status bar we need to check if the keyguard
was showing and if so make sure it is showing after the theme
change.  We also need to make sure the phone status bar is not
visible if the keyguard was showing, otherwise you see two status
bars.

Conflicts:
	packages/SystemUI/src/com/android/systemui/statusbar/phone/PhoneStatusBar.java

Change-Id: Ie52a1a0df56dbc7b7a58a9fccac6d75f3a7c8b05