Fix return to task when single task activity is brought to front
When a single task activity launched on top of home is relaunched
from another flow the return to task for the above activity is not
set correctly.

Adding a check and setting the returnto task to same as the
relaunched single task activity reurnto value ensures correct
activity is resumed.

Change-Id: Ife84aee49395fa2a1198c9ce21c75b6ef4627aef
CRs-Fixed: 830806