Themes: Allow packages to be set as non-themeable
This patch allows for specific packages to be declared as non
themeable.  Packages that are non themeable will not have theme
resources attached to their assets when their Resources object
is created.

This patch also removes isThemeable from CompatibilityInfo as this
object did not correctly reflect that the package is not themeable.

Change-Id: I7b95704f23a3ffc068a696ab61c09cee51334aca
REF: CYNGNOS-770