Let liblights adjust the brightness of LEDs while previewing it
Each device might use the information about the brightness of LEDs
in a unique way. Currently the slider to adjust the brightness of
LEDs is using an already scaled RGB color to preview the intensity
of the light. Instead of doing this, use a white colored light and
add to it the information about the brightness.

Change-Id: I07e940cb8c85a062c965bc361d73743369d402dd