Keyguard: Remove bouncer when re-registering status bar
When changing themes we restart the keyguard which causes the
StatusBarKeyguardViewManager to recreate the bouncer without
first removing the old bouncer.  This would cause a user to not
be able to correctly enter their pin since the wrong PasswordTextView
would be found via findViewById().  This patch removes the bouncer
view if mBouncer is non-null when registering the status bar.

Change-Id: I58d8f473b83836ceec9110e1d99edf09a89c98a1
REF: CYNGNOS-878