Fixed Possible Resolver Infinite Loop
There was a possibility in the case where the ApplicationSuggestions
are being used and the Resolver isn't replaced that an infinite loop
of the resolver starting itself up and finishing its current instance
would happen.  This will fix this case by no returning the Resolver
in the query if we aren't currently using a Custom Resolver.

This means that any custom resolver used needs to handle the case where
it gets returned a reference to itself when there are no other activities
to handle the intent otherwise.

Change-Id: I93a54ca13b00c33837f50480f34bc92f63b2fb27