SystemUI: bail out of recents-to-home animation when losing focus
In the case when recents is animating out, and an application happens to
immediately add a view to the window manager, the recents view would
lose its focus, but the animations would continue to run, breaking the
behavior in certain cases.

Bail out of the recents activity and override the pending transition to
be disabled to not distrubt the newly added window.

Ref: CYNGNOS-1136

Signed-off-by: Roman Birg <roman@cyngn.com>

Conflicts:
	packages/SystemUI/src/com/android/systemui/recents/RecentsActivity.java

Change-Id: Iba2f888427ec3f3eccfaca8d5de102f116237b70