SystemUI: fix "directly show bouncer" conflicting with smart unlock
If the user had "directly show pattern view" selected with a pattern
unlock for instance, then the assumption is that the lock screen always
opens to the pattern view, or the bouncer.

If smart lock is currently enabled and putting the keyguard in a
'trusted' environment, e.g. not requiring security, then we shouldn't
try to show the bouncer, because there is none. Currently when smart
lock is putting the keyguard in an insecure mode, it will basically
disable the lock screen if this option is selected. This is not the
desired result and we should still have a lock screen.

OSS Ref: BACON-4449
Signed-off-by: Roman Birg <roman@cyngn.com>

Change-Id: Ie7a2326bf06e64c171e16fec5dde1179fe975450

Conflicts:
	packages/SystemUI/src/com/android/systemui/statusbar/phone/KeyguardBouncer.java